requirements.txt
Django==1.9.5
psycopg2==2.6.1
raven==5.13.0
PyYAML==3.11
ruamel.yaml==0.11.11
ipython==4.2.0
pytz
Pillow==3.2.0
aldjemy==0.6.0
SQLAlchemy==1.0.13
python-dateutil==2.5.3
